pyproject.toml
[project]
name = "coach-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.1",
    "folium>=0.20.0",
    "garminconnect>=0.2.28",
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "plotly-express>=0.4.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "streamlit>=1.46.0",
    "streamlit-folium>=0.25.0",
]
